Remove user_role table migration with model, query and resource. Move role constants to UserResource.php and add static function for getting roles. Add first name and last name in UserResource.php fields.
